Password Strength Check with Machine Learning
This project is a password strength checker that uses machine learning to predict the strength of a password. The project is implemented in Python using the Scikit-learn library.

The project first collects a dataset of passwords and their corresponding strengths. The strengths are labeled as "weak", "medium", or "strong". The dataset is then used to train a machine learning model. The model learns to predict the strength of a password based on its length, complexity, and other factors.

Once the model is trained, it can be used to check the strength of any password. The user enters a password, and the model returns a prediction of the password's strength. The prediction is displayed to the user.
This project is a useful tool for helping users create strong passwords. It can also be used to educate users about password security.

Here are some of the key features of the project:

Uses machine learning to predict the strength of a password
Trained on a dataset of real passwords
Easy to use
Provides feedback on how to make passwords stronger